# apps/backend/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/characters": {
            "get": {
                "operationId": "CharactersController_findAll",
                "parameters": [
                    {
                        "name": "userId",
                        "required": false,
                        "in": "query",
                        "description": "Filter by user ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of characters retrieved"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Get all characters",
                "tags": [
                    "Characters"
                ]
            },
            "post": {
                "operationId": "CharactersController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCharacterDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Character created successfully"
                    },
                    "400": {
                        "description": "Invalid character data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Create a new character",
                "tags": [
                    "Characters"
                ]
            }
        },
        "/api/characters/{id}": {
            "get": {
                "operationId": "CharactersController_findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Character ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Character retrieved"
                    },
                    "404": {
                        "description": "Character not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Get character by ID",
                "tags": [
                    "Characters"
                ]
            },
            "patch": {
                "operationId": "CharactersController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Character ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCharacterDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Character updated successfully"
                    },
                    "404": {
                        "description": "Character not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Update character",
                "tags": [
                    "Characters"
                ]
            },
            "delete": {
                "operationId": "CharactersController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Character ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Character deleted successfully"
                    },
                    "404": {
                        "description": "Character not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Delete character",
                "tags": [
                    "Characters"
                ]
            }
        },
...
        "/api/inventory": {
            "get": {
                "operationId": "InventoryController_findByCharacter",
                "parameters": [
                    {
                        "name": "characterId",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Character inventory"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Get inventory for character",
                "tags": [
                    "Inventory"
                ]
            },
            "post": {
                "operationId": "InventoryController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateInventoryDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inventory item created"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Create inventory item",
                "tags": [
                    "Inventory"
                ]
            }
        },
        "/api/inventory/{id}": {
            "patch": {
                "operationId": "InventoryController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateInventoryDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inventory item updated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Update inventory item",
                "tags": [
                    "Inventory"
                ]
            },
            "delete": {
                "operationId": "InventoryController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inventory item deleted"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Delete inventory item",
                "tags": [
                    "Inventory"
                ]
            }
        },
...
        "/api/narrative": {
            "get": {
                "operationId": "NarrativeController_findBySession",
                "parameters": [
                    {
                        "name": "sessionId",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Narrative elements"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Get narrative elements for session",
                "tags": [
                    "Narrative"
                ]
            },
            "post": {
                "operationId": "NarrativeController_create",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": "Narrative element created"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Create narrative element",
                "tags": [
                    "Narrative"
                ]
            }
        },
        "/api/narrative/{id}": {
            "patch": {
                "operationId": "NarrativeController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Narrative element updated"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Update narrative element",
                "tags": [
                    "Narrative"
                ]
            },
            "delete": {
                "operationId": "NarrativeController_delete",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Narrative element deleted"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "Delete narrative element",
                "tags": [
                    "Narrative"
                ]
            }
        },
...
        "/api/gameboards/{sessionId}": {
            "get": {
                "description": "Retrieves the complete gameboard state including characters, chapters, and events",
                "operationId": "GameboardController_getGameboard",
                "parameters": [
                    {
                        "name": "sessionId",
                        "required": true,
                        "in": "path",
                        "description": "Game session ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "summary": "Get gameboard for a session",
                "tags": [
                    "gameboards"
                ]
            }
        },
        "/api/gameboards": {
            "post": {
                "description": "Initializes a gameboard for a game session",
                "operationId": "GameboardController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateGameboardDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "summary": "Create a new gameboard",
                "tags": [
                    "gameboards"
                ]
            }
        },
        "/api/gameboards/{id}": {
            "patch": {
                "description": "Updates the gameboard configuration or state",
                "operationId": "GameboardController_update",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "Gameboard ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateGameboardDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "summary": "Update gameboard state",
                "tags": [
                    "gameboards"
                ]
            }
        },
...
    },
    "info": {
        "title": "Game-Plug API",
        "description": "Call of Cthulhu 7e RPG Platform API - Complete backend specification with NestJS and OpenAPI 3.0",
        "version": "1.0.0",
        "contact": {
            "name": "Game-Plug Team",
            "url": "https://github.com/robinswood",
            "email": "contact@game-plug.dev"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
...
    "servers": [
        {
            "url": "http://localhost:4000",
            "description": "Development"
        },
        {
            "url": "http://localhost:4000",
            "description": "Production"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "CreateCharacterDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Professor John Arcane",
                        "description": "Character name"
                    },
                    "sessionId": {
                        "type": "string",
                        "example": "sess_123",
                        "description": "Session ID"
                    },
                    "occupation": {
                        "type": "string",
                        "example": "Archaeologist",
                        "description": "Character occupation"
                    },
                    "strength": {
                        "type": "number",
                        "example": 75,
                        "description": "Strength (1-99)"
                    },
                    "constitution": {
                        "type": "number",
                        "example": 70,
                        "description": "Constitution (1-99)"
                    },
                    "size": {
                        "type": "number",
                        "example": 60,
                        "description": "Size (1-99)"
                    },
                    "dexterity": {
                        "type": "number",
                        "example": 72,
                        "description": "Dexterity (1-99)"
                    },
                    "appearance": {
                        "type": "number",
                        "example": 65,
                        "description": "Appearance (1-99)"
                    },
                    "intelligence": {
                        "type": "number",
                        "example": 85,
                        "description": "Intelligence (1-99)"
                    },
                    "power": {
                        "type": "number",
                        "example": 55,
                        "description": "Power (1-99)"
                    },
                    "education": {
                        "type": "number",
                        "example": 80,
                        "description": "Education (1-99)"
                    },
                    "luck": {
                        "type": "number",
                        "example": 50,
                        "description": "Luck (1-99)"
                    },
                    "hitPoints": {
                        "type": "number",
                        "example": 10,
                        "description": "Hit points"
                    },
                    "maxHitPoints": {
                        "type": "number",
                        "example": 12,
                        "description": "Maximum hit points"
                    },
                    "sanity": {
                        "type": "number",
                        "example": 75,
                        "description": "Sanity (0-99)"
                    },
                    "maxSanity": {
                        "type": "number",
                        "example": 99,
                        "description": "Maximum sanity (0-99)"
                    },
                    "magicPoints": {
                        "type": "number",
                        "example": 5,
                        "description": "Magic points"
                    },
                    "maxMagicPoints": {
                        "type": "number",
                        "example": 10,
                        "description": "Maximum magic points"
                    },
                    "age": {
                        "type": "number",
                        "example": 45,
                        "description": "Character age"
                    },
                    "birthplace": {
                        "type": "string",
                        "example": "Cairo",
                        "description": "Birthplace"
                    },
                    "residence": {
                        "type": "string",
                        "example": "London",
                        "description": "Current residence"
                    },
                    "gender": {
                        "type": "string",
                        "example": "Male",
                        "description": "Gender"
                    },
                    "height": {
                        "type": "string",
                        "example": "5'10\"",
                        "description": "Height"
                    },
                    "build": {
                        "type": "string",
                        "example": "Athletic",
                        "description": "Build"
                    },
                    "hairColor": {
                        "type": "string",
                        "example": "Brown",
                        "description": "Hair color"
                    },
                    "eyeColor": {
                        "type": "string",
                        "example": "Blue",
                        "description": "Eye color"
                    },
                    "avatarUrl": {
                        "type": "string",
                        "example": "https://avatar.url/prof.jpg",
                        "description": "Avatar URL"
                    },
                    "avatarPrompt": {
                        "type": "string",
                        "example": "Wise academic in 1920s attire",
                        "description": "Prompt for AI avatar generation"
                    },
                    "skills": {
                        "type": "object",
                        "example": {
                            "Archaeology": 70,
                            "Library Use": 60
                        },
                        "description": "Character skills (key-value pairs)"
                    },
                    "notes": {
                        "type": "string",
                        "example": "Expert in ancient texts",
                        "description": "Player notes"
                    },
                    "money": {
                        "type": "string",
                        "example": "500.50",
                        "description": "Starting money"
                    },
                    "userId": {
                        "type": "string",
                        "example": "user_456",
                        "description": "User ID (owner)"
                    }
                },
                "required": [
                    "name",
                    "sessionId",
                    "occupation",
                    "strength",
                    "constitution",
                    "size",
                    "dexterity",
                    "appearance",
                    "intelligence",
                    "power",
                    "education",
                    "luck",
                    "hitPoints",
                    "maxHitPoints",
                    "sanity",
                    "maxSanity",
                    "magicPoints",
                    "maxMagicPoints"
                ]
            },
            "UpdateCharacterDto": {
                "type": "object",
                "properties": {}
            },
...
            "CreateInventoryDto": {
                "type": "object",
                "properties": {}
            },
            "UpdateInventoryDto": {
                "type": "object",
                "properties": {}
            },
...
            "CreateGameboardDto": {
                "type": "object",
                "properties": {
                    "sessionId": {
                        "type": "string",
                        "description": "Session ID for the gameboard"
                    },
                    "title": {
                        "type": "string",
                        "description": "Gameboard title"
                    },
                    "description": {
                        "type": "string",
                        "description": "Gameboard description"
                    },
                    "characterIds": {
                        "description": "Character IDs to include",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "state": {
                        "type": "object",
                        "description": "Initial state or configuration (JSON)"
                    }
                },
                "required": [
                    "sessionId",
                    "title"
                ]
            },
            "UpdateGameboardDto": {
                "type": "object",
                "properties": {
                    "sessionId": {
                        "type": "string",
                        "description": "Session ID for the gameboard"
                    },
                    "title": {
                        "type": "string",
                        "description": "Gameboard title"
                    },
                    "description": {
                        "type": "string",
                        "description": "Gameboard description"
                    },
                    "characterIds": {
                        "description": "Character IDs to include",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "state": {
                        "type": "object",
                        "description": "Initial state or configuration (JSON)"
                    }
                }
            },
...
        }
    }
}